move symbol sets
