+++ frontend/src/scenes/error-tracking/configuration/symbol-sets/SymbolSetUploadModal.tsx
