+++ frontend/src/scenes/error-tracking/configuration/symbol-sets/SymbolSets.tsx
