+++ frontend/src/scenes/error-tracking/configuration/symbol-sets/errorTrackingSymbolSetLogic.tsx
