+++ frontend/src/scenes/settings/types.ts
@@ -76,6 +76,7 @@ export type SettingId =
     | 'feature-flags-interface'
     | 'error-tracking-user-groups'
     | 'error-tracking-symbol-sets'
+    | 'error-tracking-alerting'
     | 'authorized-toolbar-urls'
     | 'integration-webhooks'
     | 'integration-slack'
